# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Document Intelligence & Question Extraction Service",
    "description": "\n    ## Pragati Bharati Document Intelligence & Question Extraction Service\n    \n    A production-oriented, scalable microservice to ingest examination papers and question banks\n    in PDF (digital and scanned) and Image formats, converting them into structured, machine-readable\n    question sets.\n    \n    ### Key Features:\n    * **Asynchronous Document Ingestion**: Upload large documents without blocking clients.\n    * **Automated Question Segmentation**: Detect question numbers, bodies, and options.\n    * **Cross-Page Stitching**: Intelligently handles questions continuing across page breaks.\n    * **Answer Key Matching**: Pairs answer keys from inline sections or standalone documents.\n    * **Confidence & Review Pipeline**: Flags uncertain, partial, or low-quality extractions.\n    * **Security & Auth**: JWT-based bearer authentication and user authorization isolation.\n    ",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/auth/login/json": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Login Json",
        "description": "JSON login endpoint for web frontends / mobile apps",
        "operationId": "login_json_api_v1_auth_login_json_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserLogin"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Token"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/documents/upload": {
      "post": {
        "tags": [
          "Documents"
        ],
        "summary": "Upload Document",
        "description": "Upload a PDF or Image asynchronously.\nReturns 202 Accepted immediately with document_id and tracking status.",
        "operationId": "upload_document_api_v1_documents_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_document_api_v1_documents_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentCreateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
    "/api/v1/documents/{document_id}/status": {
      "get": {
        "tags": [
          "Documents"
        ],
        "summary": "Check Document Status",
        "operationId": "check_document_status_api_v1_documents__document_id__status_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/documents/{document_id}/relate": {
      "post": {
        "tags": [
          "Documents"
        ],
        "summary": "Relate Documents",
        "description": "Associate Answer Key document with Question Paper document and trigger cross-matching.",
        "operationId": "relate_documents_api_v1_documents__document_id__relate_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentRelateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentRelateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/documents/{document_id}/questions": {
      "get": {
        "tags": [
          "Questions & Answers"
        ],
        "summary": "Get Document Questions",
        "description": "Retrieve all extracted questions for a document with optional filters and pagination",
        "operationId": "get_document_questions_api_v1_documents__document_id__questions_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          },
          {
            "name": "confidence_filter",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ConfidenceLevel"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Confidence Filter"
            }
          },
          {
            "name": "question_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/QuestionType"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Question Type"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/documents/{document_id}/answer-key": {
      "get": {
        "tags": [
          "Questions & Answers"
        ],
        "summary": "Get Document Answer Key",
        "description": "Retrieve extracted answer key entries for a document",
        "operationId": "get_document_answer_key_api_v1_documents__document_id__answer_key_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AnswerKeyEntryResponse"
                  },
                  "title": "Response Get Document Answer Key Api V1 Documents  Document Id  Answer Key Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/{document_id}/export": {
      "get": {
        "tags": [
          "Questions & Answers"
        ],
        "summary": "Export Questions Structured",
        "description": "Returns questions in the clean, system-independent schema specified in\nProblem Statement Section 7.",
        "operationId": "export_questions_structured_api_v1_documents__document_id__export_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StructuredQuestionExport"
                  },
                  "title": "Response Export Questions Structured Api V1 Documents  Document Id  Export Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/review/documents/{document_id}/queue": {
      "get": {
        "tags": [
          "Review Queue & Validation"
        ],
        "summary": "Get Review Queue",
        "description": "Retrieve questions requiring human review:\n- Low confidence (<0.8)\n- Uncertain or unmatched answers\n- Cross-page continuation splits\n- Missing options or question numbers",
        "operationId": "get_review_queue_api_v1_review_documents__document_id__queue_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReviewQueueItemResponse"
                  },
                  "title": "Response Get Review Queue Api V1 Review Documents  Document Id  Queue Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AnswerKeyEntryResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "document_id": {
            "type": "string",
            "title": "Document Id"
          },
          "question_number": {
            "type": "integer",
            "title": "Question Number"
          },
          "correct_answer": {
            "type": "string",
            "title": "Correct Answer"
          },
          "explanation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Explanation"
          },
          "source_page": {
            "type": "integer",
            "title": "Source Page"
          },
          "raw_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw Text"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "document_id",
          "question_number",
          "correct_answer",
          "source_page",
          "created_at"
        ],
        "title": "AnswerKeyEntryResponse"
      },
...
      "Body_upload_document_api_v1_documents_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "document_role": {
            "$ref": "#/components/schemas/DocumentRole",
            "default": "QUESTION_PAPER"
          },
          "target_document_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Document Id"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_document_api_v1_documents_upload_post"
      },
      "ConfidenceLevel": {
        "type": "string",
        "enum": [
          "CONFIDENT",
          "PARTIAL",
          "NEEDS_REVIEW"
        ],
        "title": "ConfidenceLevel"
      },
      "DocumentCreateResponse": {
        "properties": {
          "document_id": {
            "type": "string",
            "title": "Document Id"
          },
          "original_filename": {
            "type": "string",
            "title": "Original Filename"
          },
          "status": {
            "$ref": "#/components/schemas/DocumentStatus"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "document_id",
          "original_filename",
          "status",
          "message",
          "task_id",
          "created_at"
        ],
        "title": "DocumentCreateResponse"
      },
...
      "DocumentRelateRequest": {
        "properties": {
          "target_document_id": {
            "type": "string",
            "title": "Target Document Id"
          },
          "relationship_type": {
            "$ref": "#/components/schemas/RelationType",
            "default": "ANSWER_KEY_FOR"
          }
        },
        "type": "object",
        "required": [
          "target_document_id"
        ],
        "title": "DocumentRelateRequest"
      },
      "DocumentRelateResponse": {
        "properties": {
          "relation_id": {
            "type": "string",
            "title": "Relation Id"
          },
          "source_document_id": {
            "type": "string",
            "title": "Source Document Id"
          },
          "target_document_id": {
            "type": "string",
            "title": "Target Document Id"
          },
          "relationship_type": {
            "$ref": "#/components/schemas/RelationType"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "relation_id",
          "source_document_id",
          "target_document_id",
          "relationship_type",
          "message",
          "created_at"
        ],
        "title": "DocumentRelateResponse"
      },
      "DocumentRole": {
        "type": "string",
        "enum": [
          "QUESTION_PAPER",
          "ANSWER_KEY",
          "COMBINED",
          "SUPPLEMENTARY"
        ],
        "title": "DocumentRole"
      },
      "DocumentStatus": {
        "type": "string",
        "enum": [
          "PENDING",
          "PROCESSING",
          "COMPLETED",
          "FAILED"
        ],
        "title": "DocumentStatus"
      },
      "DocumentStatusResponse": {
        "properties": {
          "document_id": {
            "type": "string",
            "title": "Document Id"
          },
          "original_filename": {
            "type": "string",
            "title": "Original Filename"
          },
          "status": {
            "$ref": "#/components/schemas/DocumentStatus"
          },
          "progress_percent": {
            "type": "integer",
            "title": "Progress Percent"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "page_count": {
            "type": "integer",
            "title": "Page Count"
          },
          "total_questions": {
            "type": "integer",
            "title": "Total Questions"
          },
          "confident_questions": {
            "type": "integer",
            "title": "Confident Questions"
          },
          "review_required_questions": {
            "type": "integer",
            "title": "Review Required Questions"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "document_id",
          "original_filename",
          "status",
          "progress_percent",
          "page_count",
          "total_questions",
          "confident_questions",
          "review_required_questions",
          "created_at",
          "updated_at"
        ],
        "title": "DocumentStatusResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "QuestionListResponse": {
        "properties": {
          "document_id": {
            "type": "string",
            "title": "Document Id"
          },
          "total_count": {
            "type": "integer",
            "title": "Total Count"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "questions": {
            "items": {
              "$ref": "#/components/schemas/QuestionResponse"
            },
            "type": "array",
            "title": "Questions"
          }
        },
        "type": "object",
        "required": [
          "document_id",
          "total_count",
          "page",
          "page_size",
          "questions"
        ],
        "title": "QuestionListResponse"
      },
...
      "QuestionType": {
        "type": "string",
        "enum": [
          "MULTIPLE_CHOICE",
          "TRUE_FALSE",
          "SHORT_ANSWER",
          "ESSAY",
          "NUMERICAL",
          "UNKNOWN"
        ],
        "title": "QuestionType"
      },
...
      "Token": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          },
          "user": {
            "$ref": "#/components/schemas/UserResponse"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "user"
        ],
        "title": "Token"
      },
      "UserLogin": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "UserLogin"
      },
...
      "UserResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "is_admin": {
            "type": "boolean",
            "title": "Is Admin"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "username",
          "is_active",
          "is_admin",
          "created_at"
        ],
        "title": "UserResponse"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}